medical_report_analyzer/requirements.txt
streamlit
flask
google-generativeai
langchain
PyPDF2
python-dotenv
faiss-cpu
langchain-google-genai
flask-cors
langchain-community
requests